MSVC workaround for enable_if, bug fix in json
MSVC had a minor issue with the new EnableIf syntax which was remedied by,
as is always the solution in c++, more layers of templates.  See comments
in traits.hpp around EnableIfHelper for more information.

While converting json.hpp to new style EnableIf, found and fixed a small bug
not yet reported bug in overload resolution for minimal types.